Updated Ameriga Events.txt to fix CTD(issue #60)
The event to let Beornia send units to the English is broken. I suspect because the placed units have no pops to back them, which confuses the game. It crashes instantly to desktop when it happens to the AI, or on hover when it happens to the player. By only placing naval units, this crash is prevented.