Proper Caliphate
I didn't find the Caliphate to be very satisfying to play so I made it relatively unique with a special casus belli.
1. UAR is not called "The Caliphate" anymore but instead the "Arab Kingdom" (if monarchy) until you take a new "Declare Jihad" decision.
2. This decision turns you into a theocracy and gives you a permanent conquest casus belli on all non-muslim neighbours; it costs a flat 10 infamy to use. This is obviously very strong, so there are penalties:
3. Using this casus belli will also make all non-muslims in your country very militant. Winning a war with it will devastate the won territory, decimating the population and making it very militant for the next 10 years
4. Losing any war will make you lose the CB/Caliphate permanently and turn you back into a monarchy, additionally your *entire* population will get 10 militancy.

Probably needs some tweaks and while obviously strong when used with some planning, I don't think it is too overpowered and provides for a unique game of a semi-world conquest.

Feel free to test it out and decide for yourself whether it should be part of the main mod (advised country: Egypt or Iraq, just like the original UAR tag).